Add `paginationWrapper` to slack client
