+++ internal/slack/users_from_conversation.go
@@ -3,7 +3,6 @@ package slack
 import (
 	"github.com/acs-dl/slack-module-svc/internal/data"
 	"github.com/slack-go/slack"
-	"gitlab.com/distributed_lab/logan/v3"
 	"gitlab.com/distributed_lab/logan/v3/errors"
 )
 
@@ -16,54 +15,27 @@ func (c *client) GetConversationUsers(conversation data.Conversation, priority i
 	return users, nil
 }
 
-func (c *client) getUsersInConversationWrapper(params slack.GetUsersInConversationParameters, priority int) ([]string, string, error) {
-	type response struct {
-		users      []string
-		nextCursor string
-	}
-
-	wrapperFunc := func() (response, error) {
-		users, nextCursor, err := c.botClient.GetUsersInConversation(&params)
-		return response{users, nextCursor}, err
-	}
-
-	item, err := addFunctionInPQueue(
-		c.pqueues.BotPQueue,
-		wrapperFunc,
-		[]any{},
-		priority,
-	)
-	if err != nil {
-		return nil, "", errors.Wrap(err, "failed to add function in pqueue")
-	}
-
-	err = item.Response.Error
-	if err != nil {
-		return nil, "", errors.Wrap(err, "failed to get users from api", logan.F{
-			"params": params,
-		})
-	}
-
-	result, ok := item.Response.Value.(response)
-	if !ok {
-		return nil, "", errors.New("failed to convert response")
-	}
-
-	return result.users, result.nextCursor, err
-}
-
 func (c *client) getAllUsersFromConversation(conversationId string, priority int) ([]data.User, error) {
 	var users []data.User
 	cursor := ""
 
 	for {
-		params := &slack.GetUsersInConversationParameters{
+		params := slack.GetUsersInConversationParameters{
 			ChannelID: conversationId,
 			Cursor:    cursor,
 		}
-		userIDs, nextCursor, err := c.getUsersInConversationWrapper(*params, priority)
+
+		response, err := c.paginationWrapper(func() (response, error) {
+			users, nextCursor, err := c.botClient.GetUsersInConversation(&params)
+			return response{users, nextCursor}, err
+		}, priority)
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to get users in conversation")
+		}
+
+		userIDs, ok := response.payload.([]string)
+		if !ok {
+			return nil, errors.New("failed to convert response to slice of users")
 		}
 
 		for _, userID := range userIDs {
@@ -80,10 +52,10 @@ func (c *client) getAllUsersFromConversation(conversationId string, priority int
 			})
 		}
 
-		if nextCursor == "" {
+		if response.nextCursor == "" {
 			break
 		}
-		cursor = nextCursor
+		cursor = response.nextCursor
 	}
 
 	return users, nil
